openai/requirements.txt
# OpenAI Voice Assistant Requirements

# Core web framework and server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# OpenAI integration
openai>=1.0.0

# Text-to-speech
gTTS>=2.3.0

# Environment configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Logging
loguru>=0.7.0

# CORS middleware
starlette>=0.30.0

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0